Replace uint16_t with uint in drawables

+++ src/desktop.cpp
@@ -10,10 +10,10 @@
 constexpr int k_max_malloc = 16711568;
 
 dxp_desktop::dxp_desktop (
-    const int16_t x,      ///< x coordinate of the top left corner
-    const int16_t y,      ///< y coordinate of the top left corner
-    const uint16_t width, ///< width of display
-    const uint16_t height ///< height of display
+    const int16_t x,  ///< x coordinate of the top left corner
+    const int16_t y,  ///< y coordinate of the top left corner
+    const uint width, ///< width of display
+    const uint height ///< height of display
     )
     : drawable (x, y, width, height)
 {
